+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - core: update the `idr-ebr` crate to v0.3 to fix possible crash in `Context::finished()`.
 - logger: replace the `atty` crate with `IsTerminal` to fix cargo-audit warnings.
 - macros/msg: better go-to-definition via RA ([#141]).
+- macros/message: irregular warnings about `non_snake_case` by RA.
 
 [#74]: https://github.com/elfo-rs/elfo/issues/74
 [#135]: https://github.com/elfo-rs/elfo/pull/135